MODROLESKC-233: added logic to create resources for endpoints that do not exist in the mod-pubsub module but are still required to be created in the keycloak
