+++ src/main/java/org/folio/entitlement/integration/kafka/KafkaEventUtils.java
@@ -4,17 +4,12 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.folio.common.domain.model.InterfaceDescriptor;
-import org.folio.common.domain.model.ModuleDescriptor;
-import org.folio.common.domain.model.RoutingEntry;
 import org.folio.entitlement.integration.kafka.model.PermissionMappingValue;
 import org.folio.entitlement.integration.kafka.model.ResourceEvent;
 import org.folio.entitlement.integration.kafka.model.ResourceEventType;
@@ -97,19 +92,7 @@ public class KafkaEventUtils {
     return permissionMapping.containsKey(key);
   }
 
-  public static void addMissingResources(ModuleDescriptor newDescriptor) {
-    InterfaceDescriptor interfaceDescriptor = new InterfaceDescriptor();
-    interfaceDescriptor.setId("pubsub-event-handlers");
-    interfaceDescriptor.setVersion("1.1");
-    ArrayList<RoutingEntry> handlers = new ArrayList<>();
-    for (Map.Entry<String, PermissionMappingValue> mapping : permissionMapping.entrySet()) {
-      RoutingEntry routingEntry = new RoutingEntry();
-      routingEntry.setMethods(List.of(mapping.getValue().getMethod()));
-      routingEntry.setPathPattern(mapping.getValue().getEndpoint());
-      routingEntry.setPermissionsRequired(List.of(mapping.getKey()));
-      handlers.add(routingEntry);
-    }
-    interfaceDescriptor.setHandlers(handlers);
-    newDescriptor.getProvides().add(interfaceDescriptor);
+  public static Map<String, PermissionMappingValue> getPermissionMapping() {
+    return permissionMapping;
   }
 }
