+++ src/test/java/org/folio/entitlement/integration/keycloak/KeycloakUtilsTest.java
@@ -1,0 +1,22 @@
+package org.folio.entitlement.integration.keycloak;
+
+import java.util.ArrayList;
+import org.folio.common.domain.model.ModuleDescriptor;
+import org.folio.test.types.UnitTest;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+@UnitTest
+@ExtendWith(MockitoExtension.class)
+public class KeycloakUtilsTest {
+  @Test
+  void testUpdatePubSubDescriptorWithEndpoints() {
+    ModuleDescriptor newDescriptor = new ModuleDescriptor();
+    newDescriptor.setProvides(new ArrayList<>());
+    KeycloakUtils.addMissingResources(newDescriptor);
+    Assertions.assertEquals(1, newDescriptor.getProvides().size());
+    Assertions.assertEquals(3, newDescriptor.getProvides().get(0).getHandlers().size());
+  }
+}
